cli: Install executable using entry_points
This allows to have a consistent way of execution the CLI on all
platforms. On windows, an executable launcher is automatically created
after installing the package:

* https://packaging.python.org/guides/distributing-packages-using-setuptools/#console-scripts
* https://setuptools.readthedocs.io/en/latest/setuptools.html#automatic-script-creation


This commit also makes it possible for developer to modify the wordcloud/wordcloud_cli.py
script after doing an editable install and ensure the updates are considered
without having to re-install.

After downloading the source, if a user tries to directly execute "wordcloud/worcloud_cli.py",
instead of getting the error message:

  SystemError: Parent module '' not loaded, cannot perform relative import

it now shows an informative message suggesting how to run the CLI:

  To execute the CLI, instead consider running:

    wordcloud_cli --help

  or

    python -m wordcloud --help

Fixes #420