use HanLP replace Jieba
pyhanlp is One of
the most powerful natural language processing libraries in Chinese
today, and it's extremely easy to use.You can use 'PIP install pyhanlp'.
To install it,like Jieba.

Its level of identity of named entity,word segmentation was better than jieba,
and has more ways to do it.You'll save a lot of time when you use it.

And thanks to its excellent performance,
when we handle a large number of Chinese texts
We don't have to use the User-defined dictionaries.