You can use jieba or hanlp create wordcloud with chinese
Wordcloud is a very good tools, but if you want to create
Chinese wordcloud only wordcloud is not enough. The file
shows how to use wordcloud with Chinese. First, you need a
Chinese word segmentation library jieba or HanLp.You can
use 'PIP install jieba' or 'PIP install pyhanlp' or to
install it.As you can see,at the same time using wordcloud
with jieba or HanLP very convenient.While jieba is lighter,
hanlp requires more downloads, but is more powerfulHanLP's
level of identity of named entity,word segmentation was
better than jieba,and has more ways to do it.You'll save
a lot of time when you use it.